Add a more prominent add comment button
